Steps to extract and integrate BatchEuphoria
