+++ .idea/modules.xml
@@ -2,13 +2,13 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
+      <module fileurl="file://$PROJECT_DIR$/../RoddyInfra/BatchEuphoria/src/main/BatchEuphoria.iml" filepath="$PROJECT_DIR$/../RoddyInfra/BatchEuphoria/src/main/BatchEuphoria.iml" />
       <module fileurl="file://$PROJECT_DIR$/dist/plugins/DefaultPlugin/DefaultPlugin.iml" filepath="$PROJECT_DIR$/dist/plugins/DefaultPlugin/DefaultPlugin.iml" />
       <module fileurl="file://$PROJECT_DIR$/dist/plugins/PluginBase/PluginBase.iml" filepath="$PROJECT_DIR$/dist/plugins/PluginBase/PluginBase.iml" />
       <module fileurl="file://$PROJECT_DIR$/RoddyCore/RoddyCore.iml" filepath="$PROJECT_DIR$/RoddyCore/RoddyCore.iml" />
       <module fileurl="file://$PROJECT_DIR$/../RoddyInfra/RoddyToolLib/RoddyToolLib.iml" filepath="$PROJECT_DIR$/../RoddyInfra/RoddyToolLib/RoddyToolLib.iml" />
       <module fileurl="file://$PROJECT_DIR$/dist/plugins/TestPluginWithJarFile/TestPluginWithJarFile.iml" filepath="$PROJECT_DIR$/dist/plugins/TestPluginWithJarFile/TestPluginWithJarFile.iml" />
       <module fileurl="file://$PROJECT_DIR$/dist/plugins/TestPluginWithoutJarFile/TestPluginWithoutJarFile.iml" filepath="$PROJECT_DIR$/dist/plugins/TestPluginWithoutJarFile/TestPluginWithoutJarFile.iml" />
-      <module fileurl="file://$PROJECT_DIR$/RoddyCore/RoddyCore.iml" filepath="$PROJECT_DIR$/RoddyCore/RoddyCore.iml" />
     </modules>
   </component>
 </project>
