+++ RoddyCore/RoddyCore.iml
@@ -10,5 +10,6 @@
     <orderEntry type="sourceFolder" forTests="false" />
     <orderEntry type="library" name="roddyBundledLibraries" level="project" />
     <orderEntry type="module" module-name="RoddyToolLib" />
+    <orderEntry type="module" module-name="BatchEuphoria" />
   </component>
 </module>
