--- RoddyCore/src/de/dkfz/roddy/execution/jobs/JobConstants.java
@@ -1,15 +1,0 @@
-/*
- * Copyright (c) 2016 eilslabs.
- *
- * Distributed under the MIT License (license terms are at https://www.github.com/eilslabs/Roddy/LICENSE.txt).
- */
-
-package de.dkfz.roddy.execution.jobs;
-
-/**
- * Job and execution specific constant fields
- */
-public class JobConstants {
-    public static final String PRM_TOOLS_DIR = "TOOLSDIR";
-    public static final String PRM_TOOL_ID = "TOOL_ID";
-}
