--- RoddyCore/src/de/dkfz/roddy/execution/jobs/ProcessingCommands.java
@@ -1,14 +1,0 @@
-/*
- * Copyright (c) 2016 eilslabs.
- *
- * Distributed under the MIT License (license terms are at https://www.github.com/eilslabs/Roddy/LICENSE.txt).
- */
-
-package de.dkfz.roddy.execution.jobs;
-
-import java.io.Serializable;
-
-/**
- */
-public abstract class ProcessingCommands implements Serializable {
-}
