--- RoddyCore/src/de/dkfz/roddy/execution/jobs/cluster/pbs/PBSProcessingCommands.java
@@ -1,18 +1,0 @@
-/*
- * Copyright (c) 2016 eilslabs.
- *
- * Distributed under the MIT License (license terms are at https://www.github.com/eilslabs/Roddy/LICENSE.txt).
- */
-
-package de.dkfz.roddy.execution.jobs.cluster.pbs;
-
-import de.dkfz.roddy.execution.jobs.ProcessingCommands;
-
-import java.io.Serializable;
-
-/**
- * Processing commands for qsub like i.e. walltime, memory etc.
- */
-public class PBSProcessingCommands extends ProcessingCommands implements Serializable {
-
-}
